All Dodo classes now implement appropriate IDLeDOM prototypes
Most of these are still not quite correct: the old convention was to
use `foo($value = NULL)` methods for attribute access instead of
`getFoo()` and `setFoo()`, and I haven't found and corrected all of
these.  But enough works that the (very basic) test case passes.  The
rest can be dealt with file-by-file; the phan errors suppressed at the
top of each file are a good guide to what remains to be fixed.

Change-Id: I4decb87b0a4aec167d23d507277f3625b02a0ebc