Implement Document::getElementById and factor out ContainerNode/Leaf
Fix the implementation of the _index_to_element and _id_to_element maps
in Document which power Document::getElementById().  This required
auditing the way that mutation is tracked (in order to keep these maps
up to date).  In the process, factored out child management into
ContainerNode and Leaf superclasses (like domino does) in order to
avoid using memory for children lists on leaf nodes.

Change-Id: I2969d8e73f47a046ab7807542c1f034b5fa03d9f