Delint WhatWG; rewrote XML name validity checking
Phan was raising quite correct issues with the invalidity of the
regular expressions that were being used for XML name validity
checking.  They were borrowed from the domino port, and the difference
in regexp engines and UTF-16 vs UTF-8 made most of what domino was
doing wrt surrogates unnecessary in PHP land.  Go back to the original
spec and simplify the productions, getting rid of the phan complaints
in the process.

Change-Id: I7132f9cb85881124f41100892df47965c3f93e1c