Clean up WPT harness helper Common.php
These methods come from common.js in the WPT test harness.  Make the
methods static and namespace them so they don't pollute the namespace
unnecessarily and it's clear where to find their definitions.  Also
restore the function order in Common.php to match the order in
common.js so that it's easier to relate the PHP implementation to the
JS implementation, especially if upstream were to make changes.  Fix a
number of bugs in the port and delint the code so that it passes phan
and phpcs.

Change-Id: I5d12a48a6519acb27ed66e7670df27a584a257d4