Switch to precise dependency on IDLeDOM
As the IDLeDOM package is still new, even "safe" minor version bumps to
IDLeDOM can cause Dodo breakage because Dodo is not using IDLeDOM in the
recommended way (ie, not extending all necessary Helper classes) etc.

So use a precise dependency for now to avoid unexpected CI breakage when
new versions of IDLeDOM are released.

Change-Id: If7469b9759308b6c3b8037e3b7a91019f6ca0c10