Create FakeElement class to reduce code duplication in Document/DocumentFragment
The ParentNode and NonElementParentNode mixins in the DOM spec contain
methods which treat Document and DocumentFragment "as if they were an
Element".  Instead of adding special code to methods in Element to handle
those cases, just add a special FakeElement class and invoke the methods
on the FakeElement directly.  This reduces code duplication and special
cases.

Implement DocumentFragment::getElementById() using this strategy, in
addition to porting the querySelector* and getElementsBy* methods.

Change-Id: Ifa7886e5a113166128459fd3317cea51606e2d6a