Ensure Dodo does not use magic properties internally
This disables the phan `read_magic_property_annotations` configuration
option to ensure that phan checks Dodo code without using magic properties.
For efficiency, all phan internal code should use direct access to the
getter or setter method as needed.

Test code is still allowed to use getters and setters, since much of it
is auto-generated from javascript code.

The key piece here is actually enabling
@phan-forbid-undeclared-magic-properties in the Dodo classes to ensure
that property accesses are checked.  This phan annotation was already
present in the IDLeDOM interfaces, but phan only checks the direct
class inheritance tree for this property, not the implemented
interfaces.

Change-Id: I7ee2d388101517acec73489c528582d9f29ea7d6