Use only one property whether a node's children is a linked list or an array
For performance almost all node will store their children as a linked list.
Avoid wasting an extra property for an array which is rarely used; instead
use a single property for both.  As a side benefit, this plays nicer with
phan, which can tell by the instanceof checks whether we're currently in
"array mode" or "linked list mode".

Change-Id: Ib594a1207e4c20bb5864324598b90256d40dfb59